Refactor DMI processing to make it harder to crash.
A lot of fragility in this code comes from having the field parsing
for the various option types blatantly ignore the length of the data
section of the header, and when it is shorter than expected they
wander off into trying to interpret embedded strings improperly or the
next DMI info structure or something else similarly insane.

This refactors the dmiHeader handling code to make all bytes they
should not be messing with into zeros instead, which should mostly do
The Right Thing.